1341-SPLIT-A-STRING-IN-BALANCED-STRINGS Stats: Time: 0 ms (100%), Space: 38.6 MB (27.27%) - LeetHub

+++ 1341-split-a-string-in-balanced-strings/1341-split-a-string-in-balanced-strings.cs
@@ -1,9 +1,9 @@
 public class Solution {
     public int BalancedStringSplit(string s) {
         ushort result = 0;
-        ushort lCount = 0;
-        ushort rCount = 0;
-        for(int i = 0; i < s.Length; i++)
+        byte lCount = 0;
+        byte rCount = 0;
+        for(ushort i = 0; i < s.Length; i++)
         {
             if(s[i] == 'R')
             {
